clase

+++ DAM 3er_año_1º-2025-26/Programación/Proyecto Programacion - Oscar Ubeda/src/main/java/org/example/Vista/CrearProyecto.java
@@ -86,7 +86,7 @@ public class CrearProyecto extends JFrame {
                 }
 
                 String nombreProyecto = Objects.requireNonNull(proyectosEnum).toString();
-                double presupuestoProyecto = (double) Objects.requireNonNull(presupuestoEnum).getMin();
+                double presupuestoProyecto = Objects.requireNonNull(presupuestoEnum).getMin();
 
                 Date fechaActual = new Date();
